(golint) rename package to tests
