+++ tests/utils_test.go
@@ -1,4 +1,4 @@
-package acceptance_tests
+package tests
 
 import (
 	"bytes"
